added utils function for getting concert image for concert card
