+++ client/Components/Card/TempCard.js
@@ -16,6 +16,7 @@ import {
 import axios from 'axios';
 
 import { GlobalState } from '../../contexts/Store';
+import { getDateInStringFormat, getConcertImage } from './utils';
 
 const useStyles = makeStyles({
     root: {
@@ -39,6 +40,10 @@ const useStyles = makeStyles({
     clear: {
         color: 'red',
     },
+    link: {
+        marginTop: '1rem', // add button instead
+        color: 'white',
+    },
 });
 
 export default function TempCard({ concertData, isAttending }) {
@@ -61,28 +66,7 @@ export default function TempCard({ concertData, isAttending }) {
         await getUserData();
     };
 
-    const getDateInStringFormat = (date) => {
-        const year = date.slice(0, 4);
-        const day = date.slice(-2);
-        let month = date.slice(5, 7);
-        if (month === '01') month = 'January';
-        if (month === '02') month = 'February';
-        if (month === '03') month = 'March';
-        if (month === '04') month = 'April';
-        if (month === '05') month = 'May';
-        if (month === '06') month = 'June';
-        if (month === '07') month = 'July';
-        if (month === '08') month = 'August';
-        if (month === '09') month = 'September';
-        if (month === '10') month = 'October';
-        if (month === '11') month = 'November';
-        if (month === '12') month = 'December';
-        return `${month} ${day}, ${year}`;
-    };
-
     console.log('concertData', concertData);
-
-    const eventImage = concertData.images[3].url; // write a function that returns a 16:9 aspect ratio or default image url
 
     return (
         Object.keys(concertData).length > 0 && (
@@ -90,7 +74,7 @@ export default function TempCard({ concertData, isAttending }) {
                 <CardActionArea>
                     <CardMedia
                         className={classes.media}
-                        image={eventImage}
+                        image={getConcertImage(concertData)}
                         title="glass"
                     />
                     <CardContent>
@@ -112,8 +96,9 @@ export default function TempCard({ concertData, isAttending }) {
                                 onClick={() => {
                                     setCurrConcert(concertData);
                                 }}
+                                className={classes.link}
                             >
-                                Click here to get to single concert page
+                                Click here to see event information
                             </Link>
                         </Typography>
                     </CardContent>
