+++ client/Components/Sidebar/Sidebar.scss
@@ -14,7 +14,7 @@
 
 .nav-menu {
     background-color: rgba(20, 20, 30, 0.8);
-    width: 18rem;
+    width: 20rem;
     height: 93vh;
     display: flex;
     justify-content: center;
